Rerwite syntex to use the new MacroResolver
